% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounce-monotone-boundary.R
\name{bounce_monotone}
\alias{bounce_monotone}
\title{Bounce away from monotone boundary}
\usage{
bounce_monotone(gam, Y, Xo, poly_basis, oracle_fun, region, basis_cv, control)
}
\arguments{
\item{gam}{gamma iterate (should be on boundary of monotonicity)}

\item{Y}{Y data}

\item{Xo}{Orthonormal X design matrix}

\item{poly_basis}{Polynomial basis for gamma}

\item{oracle_fun}{Monotonicity oracle function}

\item{region}{Region of monotonicity constraint}

\item{basis_cv}{converter function for polynomial basis}

\item{control}{control list created by \code{cols_control}}
}
\value{
new gamma estimate or original if no better soluton found
}
\description{
This function helps if COLS gets stuck in flat spots of the constraint boundary.
It will bounce the starting point towards an inner point of the constrained region.
}
\examples{
#To do

}
