% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/bootstrap-mixed-effect-em-constrained-polynomial-model.R
\name{case_bootstrap_constrained_lmm_em}
\alias{case_bootstrap_constrained_lmm_em}
\title{Bootstrap for fitting polynomial constrained mixed effect model with monotonic mean and (potentially) monotonic random effects.}
\usage{
case_bootstrap_constrained_lmm_em(model, N, maxit = 25, tol = 0.01, start,
  verbose = F, save_steps = F, check_up = F, verbose_boot = verbose,
  parallel = F, cores = 1)
}
\arguments{
\item{model}{model specification made by gcreg:::make_em_model_specs().}

\item{N}{}

\item{maxit}{maximum number of iterations for each resample.}

\item{tol}{relative tolerance for convergence of parameter values for each resample.}

\item{start}{start values for parameters for each resample. Defaults to NULL.}

\item{verbose}{output current iteration and quasi-log-likelihood value in each resample.}

\item{save_steps}{should each em step be saved and return in a list?}

\item{check_up}{should the algorithm check if the quasi-log-likelihood is improving each iteration?}

\item{verbose_boot}{Output when a bootstrap resample is finished.}

\item{parallel}{Execute in parallel. This will turn off verbose settings. ONLY IMPLEMENTED FOR LINUX/MAC}

\item{cores}{Number of cores to use.}
}
\value{
list including estimated mixed effects model.
}
\description{
EXPIREMENTAL: For use with block-random effects (non-crossed), e.g. subject-specific random effects.
This function is under development and may change without warning. See the mixed effects vignette for usage details.
}
\details{
Uses case bootstrapping, where a sample of groups is chosen with replacement and the new estimates are used.
}
\keyword{internal}
