% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-creg.R
\name{coef.creg}
\alias{coef.creg}
\title{Methods for class "creg"}
\usage{
\method{coef}{creg}(object, ...)
}
\arguments{
\item{object}{The constrained model}

\item{...}{Additional arguments to be passed to function}
}
\description{
Methods for class "creg"
}
