% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols-controls.R
\name{cols_control}
\alias{cols_control}
\title{Create controls attributes which tune the COLS algorithm.}
\usage{
cols_control(method = "simple", maxit = 200, tol = 1e-05,
  step_start = 0.7, step_increment = 0.05)
}
\arguments{
\item{method}{"simple" or "best-step".}

\item{maxit}{maximum number of iterations.}

\item{tol}{tolerance dictating when convergence is reached.}

\item{step_start}{the initial step size taken.}

\item{step_increment}{how step size increases over iterations.}
}
\value{
a list of control parameters.
}
\description{
Create controls attributes which tune the COLS algorithm.
}
\examples{
TO DO
}
