% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-oracle.R
\name{combine_oracles}
\alias{combine_oracles}
\title{WIP: Methods for manipulating oracle functions}
\usage{
combine_oracles(...)
}
\arguments{
\item{...}{oracle function(s)}

\item{f_oracle}{oracle function}

\item{x}{data}

\item{f_scale}{scaling function}
}
\value{
function that returns TRUE if point satisfies oracle
}
\description{
scale_oracle(): Specify one of data or scaling function
}
\examples{
#To do

}
