% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/mixed-effect-mcem-constrained-polynomial-model.R
\name{constrained_lmm_mcem}
\alias{constrained_lmm_mcem}
\title{Fit monotone-constrained polynomial LMM with an MCEM algorithm}
\usage{
constrained_lmm_mcem(model, maxit = 200, tol, start = NULL, verbose = F,
  save_steps = F, check_up = F, parallel = F, cores = 1,
  ni_ssize = 20000)
}
\arguments{
\item{model}{model specification made by \code{\link{make_em_model_specs}}.}

\item{maxit}{maximum number of iterations.}

\item{tol}{relative tolerance for convergence of parameter values.}

\item{start}{start values for parameters. Defaults to NULL.}

\item{verbose}{output current iteration and quasi-log-likelihood value.}

\item{save_steps}{should each em step be saved and return in a list?}

\item{check_up}{should the algorithm check if the quasi-log-likelihood is improving each iteration?}

\item{parallel}{Execute Monte Carlo expectations in parallel. This will turn off verbose settings for expectations. ONLY IMPLEMENTED FOR LINUX/MAC}

\item{cores}{Number of cores to use.}

\item{ni_ssize}{Number of realisations to used for numerical integration.}
}
\value{
list including estimated mixed effects model.
}
\description{
EXPIREMENTAL: Fit polynomial constrained mixed effect model with BOTH monotonic mean and monotonic random effects.
For use with block-random effects (non-crossed), e.g. subject-specific random effects. Uses a Monte Carlo expectation step.
}
\details{
The interface and default actions of this function are under development and may change without warning. See the mixed effects vignette for usage details.
}
\keyword{internal}
