% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-creg.R
\name{fitted.creg}
\alias{fitted.creg}
\title{Fitted values}
\usage{
\method{fitted}{creg}(object, ...)
}
\arguments{
\item{object}{The constrained model}

\item{...}{Additional arguments to be passed to function}
}
\description{
Fitted values
}
