% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-creg.R
\name{formula.creg}
\alias{formula.creg}
\title{Regression coefficients}
\usage{
\method{formula}{creg}(x, ...)
}
\arguments{
\item{x}{The constrained model}

\item{...}{Additional arguments to be passed to function}
}
\description{
Regression coefficients
}
