% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcreg.R
\docType{package}
\name{gcreg}
\alias{gcreg}
\alias{gcreg-package}
\title{gcreg: A package for fitting constrained regression models}
\description{
Currently under development. See \url{https://github.com/bonStats/gcreg} for updates.
}
\details{
Function \code{\link{cpm}} fits constrained polynomial models with fixed effects.

Hidden functions \code{\link{constrained_lmm_em}} and \code{\link{constrained_lmm_mcem}} fit constrained polynomial models with mixed effects.

Hidden function \code{\link{case_bootstrap_constrained_lmm_em}} implements case bootstrapping for standard errors for constrained mixed effects models.

Access hidden functions with \code{gcreg:::} prefix, e.g. \code{?gcreg:::constrained_lmm_em}.
}
