% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial-basis-converter.R
\name{gen_poly_basis_converters}
\alias{gen_poly_basis_converters}
\title{Generate functions to convert regression coefficients between orthonormal and standard (monomial) polynomial bases.}
\usage{
gen_poly_basis_converters(poly_basis)
}
\arguments{
\item{poly_basis}{polylist defining the orthonormal basis to be used in conversions}
}
\value{
list of length 2 with functions having names 'to_mono' (conversion to monomial basis) and 'to_ortho' (conversion to orthonormal basis)
}
\description{
Generate functions to convert regression coefficients between orthonormal and standard (monomial) polynomial bases.
}
