% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-data.R
\name{gen_scale_data_funs}
\alias{gen_scale_data_funs}
\title{Create function to scale and unscale data.}
\usage{
gen_scale_data_funs(x)
}
\arguments{
\item{x}{vector of data.}
}
\description{
Scale x variables to be between -1 and 1 which improves stability of algorithms working with polynomials
}
\examples{
x <- rnorm(100, mean = 3)
sc <- gen_scale_data_funs(x)
x_scaled <- sc$scale(x)
all(abs(sc$unscale(x_scaled) - x) < 1e-10)

}
\keyword{internal}
