% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-monotone.R
\name{is_monotone}
\alias{is_monotone}
\title{Check if polynomial is monotone over desired region.}
\usage{
is_monotone(p, region = c(-Inf, Inf), type = NULL, EPS = 1e-05)
}
\arguments{
\item{p}{polynomial}

\item{region}{numeric is polynomial monotone over this region.}

\item{type}{character one of "increasing", "decreasing" or NULL}
}
\value{
TRUE if polynomial is monotone within region, FALSE otherwise.
}
\description{
Check if polynomial is monotone over desired region.
}
\examples{
pol <- polynomial(c(1,4,5,2))
is_monotone(pol, region = c(1,5))
}
