% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear-distance-monotone-boundary.R
\name{linear_dist_to_monotone_boundary}
\alias{linear_dist_to_monotone_boundary}
\title{Calculate the linear distance to the monotone boundary}
\usage{
linear_dist_to_monotone_boundary(gam, region, basis_cv, EPS = 1e-06)
}
\arguments{
\item{gam}{gamma iterate (should be on boundary of monotonicity)}

\item{region}{Region of monotonicity constraint}

\item{basis_cv}{converter function for polynomial basis}

\item{EPS}{Precision for root finding}
}
\value{
distance to the monotone boundary (intercept of derivative - in monomial basis)
}
\description{
Calculate the linear distance to the monotone boundary
}
\examples{
#To do
}
