% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-oracle.R
\name{make_oracle}
\alias{make_oracle}
\title{WIP: Methods for manipulating oracle functions}
\usage{
make_oracle(oracle, region = c(-Inf, Inf))
}
\arguments{
\item{oracle}{oracle function, returns TRUE if point is in constrained region, FALSE otherwise.}

\item{region}{region over which constraint is applied.}
}
\value{
function that returns TRUE if point satisfies oracle, FALSE otherwise.
}
\description{
WIP: Methods for manipulating oracle functions
}
\examples{
#To do

}
