% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrained-orthonormal-least-squares.R
\name{optim_cols}
\alias{optim_cols}
\title{The COLS algorithm.}
\usage{
optim_cols(par, Y, X, oracle_fun, control = cols_control(...), aim_gamma,
  save_gamma = F, ...)
}
\arguments{
\item{par}{initial starting point of algorithm.}

\item{Y}{(scaled) observation matrix.}

\item{X}{orthonormalised design matrix.}

\item{oracle_fun}{function that determines if a given point is inside the constrained space.}

\item{control}{list of controls generated by cols_control().}

\item{aim_gamma}{The unconstrained solution to descend towards, on the orthonormal scale, computed as \code{crossprod(X, Y)} if not provided.}

\item{save_gamma}{Save and return a list of the gamma iterates.}

\item{...}{additional parameters passed to control if not specified.}
}
\value{
constrained minimum solution.
}
\description{
Provided for advanced users. Using cpm() is recommended as main interface.
}
\keyword{internal}
