% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly-basis-design-matrix.R
\name{poly_basis_design_matrix}
\alias{poly_basis_design_matrix}
\title{Create design matrix from polynomial basis.}
\usage{
poly_basis_design_matrix(poly_basis, x)
}
\arguments{
\item{poly_basis}{polylist defining the (non-monomial) basis to convert from.}

\item{x}{data for design matrix}
}
\value{
Design matrix for polynomial regression in \code{ploy_basis}
}
\description{
Create design matrix from polynomial basis.
}
