% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-creg.R
\name{predict.creg}
\alias{predict.creg}
\title{Predicted (or fitted) values}
\usage{
\method{predict}{creg}(object, newdata, ...)
}
\arguments{
\item{object}{The constrained model}

\item{newdata}{Data to be provided to predict()}

\item{...}{Additional arguments to be passed to function}
}
\description{
Predicted (or fitted) values
}
