% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-creg.R
\name{print.creg}
\alias{print.creg}
\title{Print regression model}
\usage{
\method{print}{creg}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{The constrained model}

\item{digits}{Minimal number of significant digits for print()}

\item{...}{Additional arguments to be passed to function}
}
\description{
Print regression model
}
