% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-creg.R
\name{residuals.creg}
\alias{residuals.creg}
\title{Regression residuals}
\usage{
\method{residuals}{creg}(object, ...)
}
\arguments{
\item{object}{The constrained model}

\item{...}{Additional arguments to be passed to function}
}
\description{
Regression residuals
}
