% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-parameters.R
\name{transform_beta}
\alias{transform_beta}
\title{Transform beta to gamma polynomial coefficients}
\usage{
transform_beta(b, sc_x_f, sc_y_f, cv)
}
\arguments{
\item{b}{beta, standard polynomial coefficients w.r.t. raw data.}

\item{sc_x_f}{scale functions for x data created by \code{\link{gen_scale_data_funs}}.}

\item{sc_y_f}{scale functions for y data created by \code{\link{gen_scale_data_funs}}.}

\item{cv}{orthogonal basis converter functions created by \code{\link{gen_poly_basis_converters}}.}

\item{g}{gamma, orthogonalised polynomial coefficients w.r.t. scaled data (\eqn{-1 < x, y < 1}).}
}
\description{
beta = standard polynomial coefficients w.r.t. raw data.
}
\details{
gamma = orthogonalised polynomial coefficients w.r.t. scaled data (\eqn{-1 < x, y < 1}).
}
