% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-parameters.R
\name{untransform_gam}
\alias{untransform_gam}
\title{Transform gamma to beta polynomial coefficients}
\usage{
untransform_gam(g, sc_x_f, sc_y_f, cv)
}
\arguments{
\item{g}{gamma, orthogonalised polynomial coefficients w.r.t. scaled data (\eqn{-1 < x, y < 1}).}

\item{sc_x_f}{scale functions for x data created by \code{\link{gen_scale_data_funs}}.}

\item{sc_y_f}{scale functions for y data created by \code{\link{gen_scale_data_funs}}.}

\item{cv}{orthogonal basis converter functions created by \code{\link{gen_poly_basis_converters}}.}
}
\description{
beta (\code{b}) is standard polynomial coefficients w.r.t. raw data.
gamma (\code{g}) is the orthogonalised polynomial coefficients w.r.t. scaled data (\eqn{-1 < x, y < 1}).
See \code{\link{transform_beta}} for inverse transformation.
}
