% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cols-controls.R
\name{cols_control}
\alias{cols_control}
\title{Create controls attributes which tune the COLS algorithm.}
\usage{
cols_control(method = "best-step", maxit = 200, tol = 1e-05,
  step_start = 0.7, step_increment = 0.05, maxit_bounces = 2,
  maxit_bumps = 5, tol_bounce = tol)
}
\arguments{
\item{method}{one of "best-step", "once-best-step", "up-walk", "down-walk", "avoid-boundary"}

\item{maxit}{maximum number of iterations.}

\item{tol}{tolerance dictating when convergence is reached.}

\item{step_start}{the initial step size taken.}

\item{step_increment}{how step size increases over iterations.}

\item{maxit_bounces}{Maximum number of bounces away from the solution.}

\item{maxit_bumps}{Maximum number of bumps to test for each solution.}

\item{tol_bounce}{tolerance level bounce iterations.}
}
\value{
a list of control parameters.
}
\description{
Create controls attributes which tune the COLS algorithm.
}
\examples{
cols_control(maxit = 100)

}
