% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-oracle.R
\name{combine_oracles}
\alias{combine_oracles}
\title{WIP: Combine oracle functions to be used in contrained estimation procedure}
\usage{
combine_oracles(...)
}
\arguments{
\item{...}{oracle functions to be combined}
}
\value{
function that returns TRUE if point satisfies oracle
}
\description{
WIP: Combine oracle functions to be used in contrained estimation procedure
}
\examples{
#To do

}
