% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcreg.R
\docType{package}
\name{gcreg}
\alias{gcreg}
\alias{gcreg-package}
\title{gcreg: A package for fitting constrained regression models}
\description{
Currently under development. Function cpm() fits constrained polynomial models.
}
