% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line-search.R
\name{gen_line_search}
\alias{gen_line_search}
\title{Search for boundary of permissible region between current value and desired value.}
\usage{
gen_line_search(oracle_fun, EPS = 1e-05, maxit = 200, ...)
}
\arguments{
\item{oracle_fun}{oracle function to determine membership to constrained space.}

\item{EPS}{precision of line search algorithm}

\item{maxit}{maximum number of iterations}

\item{...}{extra arguments. Currently ignored.}
}
\value{
Function to conduct line search with specified parameters.
}
\description{
Search for boundary of permissible region between current value and desired value.
}
