% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/discrete-orthonormal-polynomial-basis-generator.R
\name{make_disc_orthonormal_basis}
\alias{make_disc_orthonormal_basis}
\title{Create discrete (data-based) orthonormal polynom::polynomial basis}
\usage{
make_disc_orthonormal_basis(x, deg)
}
\arguments{
\item{x}{data to generat}

\item{deg}{Highest degree of polynomials.}
}
\value{
polylist polynomials from polynom package defining orthonormal basis
}
\description{
Create discrete (data-based) orthonormal polynom::polynomial basis
}
\examples{
To do
t(X) \%*\% X = diag()
crossprod(X) == diag()
}
