% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-em-model-specifications.R
\name{make_em_model_specs}
\alias{make_em_model_specs}
\title{Create list with model specifications for polynomial constrained mixed effect model with monotonic mean and (potentially) monotonic random effects.}
\usage{
make_em_model_specs(formula, data, p_degree, r_degree, r_constrained,
  mcontr_region = c(-1, 1), rcontr_region = c(-1, 1), group_name = "grp")
}
\arguments{
\item{formula}{Usage: y ~ x, replaced by actual names.}

\item{data}{data.frame containing y and x.}

\item{p_degree}{degree of mean polynomial.}

\item{r_degree}{degree of polynomial random effects.}

\item{r_constrained}{should the polynomial random effects be constrained? Only implemented for r_degree= 0, 1.}

\item{mcontr_region}{Over what region should monotonicity apply to the mean polynomial?}

\item{rcontr_region}{Over what region should monotonicity apply to the random effects polynomials?}

\item{group_name}{character column name of group in data.}
}
\value{
list to be passed to constrained_lmm_em() for fitting model.
}
\description{
EXPIREMENTAL: See vignette for usage details.
}
\keyword{internal}
