% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-oracle.R
\name{make_oracle}
\alias{make_oracle}
\title{Create oracle function for constrained polynomial fitting}
\usage{
make_oracle(oracle, region = c(-Inf, Inf))
}
\arguments{
\item{oracle}{oracle function, returns TRUE if point is in constrained region, FALSE otherwise.}

\item{region}{region over which constraint is applied.}
}
\value{
function that returns TRUE if point satisfies oracle, FALSE otherwise.
}
\description{
For use with \code{\link{cpm}}. See getting started vignette for details.
}
\examples{
\dontrun{
constraint_oracle <- make_oracle( function(b) { b[1,] > 1 })
# In general, variable b should be treated as a matrix.
}

}
