% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrained-orthonormal-least-squares.R
\name{optim_COLS}
\alias{optim_COLS}
\title{The COLS algorithm.}
\usage{
optim_COLS(par, Y, X, oracle_fun, control = cols_control(...), ...)
}
\arguments{
\item{par}{initial starting point of algorithm.}

\item{Y}{(scaled) observation matrix.}

\item{X}{orthonormalised design matrix.}

\item{oracle_fun}{function that determines if a given point is inside the constrained space.}

\item{control}{list of controls generated by cols_control().}

\item{...}{additional parameters passed to control if not specified.}
}
\value{
contrained minimum solution.
}
\description{
The COLS algorithm.
}
\examples{
#TO DO
}
